+++ bika/lims/browser/viewlets/analysisrequest.py
@@ -48,11 +48,20 @@ class PrimaryAnalysisRequestViewlet(ViewletBase):
     def get_partitions(self):
         """Returns whether this viewlet is visible or not
         """
+        partitions = []
+
         # If current user is a client contact, rely on Setup's ShowPartitions
-        if api.get_current_client():
+        client = api.get_current_client()
+        if client:
             if not api.get_setup().getShowPartitions():
-                return []
-        return self.context.getDescendants()
+                return partitions
+
+        partitions = self.context.getDescendants()
+        if client:
+            # Do not display partitions for Internal use
+            return filter(lambda part: not part.getInternalUse(), partitions)
+
+        return partitions
 
     def get_primary_bound_fields(self):
         """Returns a list with the names of the fields the current user has