+++ bika/lims/interfaces/__init__.py
@@ -410,6 +410,11 @@ class IAnalysisSpecs(Interface):
     """
 
 
+class IDynamicResultsRange(Interface):
+    """Marker interface for Dynamic Result Range
+    """
+
+
 class IAnalysisProfile(Interface):
     """Marker interface for an Analysis Profile
     """