+++ CHANGES.rst
@@ -10,6 +10,8 @@ Changelog
 - #1506 Specification non-compliant viewlet in Sample
 - #1506 Sample results ranges out-of-date viewlet in Sample
 - #1506 Warn icon in analyses when range is not compliant with Specification
+- #1492 Dynamic Analysis Specifications
+- #1507 Support for semi-colon character separator in CCEmails field
 - #1499 Moved navigation portlet into core 
 - #1498 Moved all viewlets from senaite.lims to senaite.core
 - #1505 Display partition link in analyses listing
@@ -22,6 +24,7 @@ Changelog
 
 **Changed**
 
+- #1508 Do not try to render InstrumentQCFailuresViewlet to non-lab personnel
 - #1495 Better Remarks handling and display
 - #1502 Improved DateTime Widget
 - #1490 Support Dexterity Behavior Fields in API
@@ -39,6 +42,9 @@ Changelog
 - #1506 Changes via manage results don't get applied to partitions
 - #1506 Fix recursion error when getting dependencies through Calculation
 - #1506 setter from ARAnalysisField does no longer return values
+- #1512 QC Analyses listing appears empty in Sample view
+- #1510 Error when viewing a Sample w/o Batch as client contact
+- #1511 Links to partitions for Internal Use are displayed in partitions viewlet
 - #1505 Manage Analyses Form re-applies partitioned Analyses back to the Root
 - #1503 Avoid duplicate CSS IDs in multi-column Add form
 - #1501 Fix Attribute Error in Reference Sample Popup